Ignore final '.hdr', '.cfl', '.'  (closes #3)
For filenames specified on the command line, the ending of '.hdr', '.cfl'
and  '.' are now ignored.

Ignoring '.' is useful for tab-completion in the shell, since it will by
default complete to the cfl name with a final dot '.'.

Ignoring '.hdr' and '.cfl' is useful since now this viewer can be set as the
default program to open these files from a file manager.