Sync connection to new function, view_s hidden
The implementation details are moved from the header back to the
implementation file.
The details of connecting sync to different view_s's is
moved to a new function, which is now called both in main() and in
window_clone(). This change is necessary since the old implementation in main
relied on the implementation details of view_s, which are now hidden.