cfl2png: fewer dependencies
We move some of the code around so that cfl2png no longer depends on
GTK+, since it really is an unneeded dependency.

Ugly part: draw now contains some functions that should rather belong in
some kind of general "utils" file or something similar, since they are
not related to drawing.