fix segfault on toggle_sync in newer GTK
With newer versions of GTK+ (at least with 3.22), a segfault is
triggered on changing sync state. The reason for that is that the signal
is called with 2 arguments, but the third argument is used as
struct view_s. According to
	 https://developer.gnome.org/gtk3/stable/GtkToggleButton.html#GtkToggleButton-toggled
the handler is called as
void
user_function (GtkToggleButton *togglebutton,
               gpointer         user_data)
so the first arguemnt is the button (which is unused in view) and the
second is the data we want.

This works on GTK+3 v3.22 and v3.14 at least. I do not know why it
worked before v3.22.