change status bar when changing pos
Previously, the status bar was only updated when pressing the secondary
mouse button, but not when changing the position through one of the spin
buttons.

Now, not only the crosshair but also the status bar is updated when the
position is changed through a spin button.

The `clear_status_bar` and `update_status_bar` functions are moved
around to avoid forward declarations.