BUGFIX: screen2pos: correct calculation of x and y
This commit corrects the calculation of the x and y position in
`screen2pos()` for flipped coordinates.

Previously, the values calculated in `screen2pos()` were truncated to
integers later (either in passing them to `set_position()` or in
`update_status_bar()`). Now, they are explicitly rounded in
`screen2pos()`.

Furthermore, this commits removes the truncation in
`update_status_bar()` and marks the `pos` argument as const where
appropriate.